Add slug generation to UpdateCategory method
A new line was added to the `UpdateCategory` method in the
`CategoryService` class to generate a slug using the
`SlugHelper.GenerateSlug` method. This assigns a slug to the
`Slug` property of the `category` object based on its name
and ID.

Additionally, the closing brace of the `CategoryService`
class was removed, which may indicate structural changes or
an accidental deletion.